+++ RELEASES.md
@@ -15,6 +15,7 @@ vNext
 - Improved: Handle errors gracefully instead of panic
 - Improved: Pretty print (line break) the class list output
 - Improved: Move baseUrl, username, password and system flags from root to sub commands
+- Improved: Pretty print (line break) the request when an error occurs
 - Fixed: Backslash meant to escape quotes is contained in description of the class command
 
 v0.3.0 (2018-04-29)